# docs/openapi.yaml
openapi: 3.1.0
info:
  title: HIMATE Control Plane API
  version: 0.7.0-start-21
  description: >
    Authenticated REST contract for the START-01 through START-21 HIMATE
    control-plane surface. Partner business databases remain isolated behind
    connector/API contracts; provider deployment is isolated behind the Runtime
    adapter service; encrypted backup/recovery is isolated behind Backups.
servers:
  - url: /
...
paths:
...
  /api/v1/backups:
    get:
      tags: [Backups]
      summary: List restore points
      parameters:
        - {name: partner_id, in: query, schema: {type: string}}
        - {name: status, in: query, schema: {type: string}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 50}}
      responses:
        "200":
          description: Restore points
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: {type: array, items: {$ref: "#/components/schemas/RestorePoint"}}
                  count: {type: integer}
    post:
      tags: [Backups]
      summary: Queue an encrypted offsite restore point
      description: >
        Captures the isolated partner database, media namespace and redacted
        configuration. A successful backup automatically queues a restore test.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [partner_id]
              properties:
                partner_id: {type: string}
      responses:
        "202":
          description: Restore point queued
          content:
            application/json:
              schema: {$ref: "#/components/schemas/RestorePoint"}
        "409": {description: Backup disabled or already queued/running}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: himate_session
    connectorBearer:
      type: http
      scheme: bearer
      bearerFormat: HIMATE Connector Credential
...
  schemas:
...
    RestorePoint:
      type: object
      properties:
        id: {type: string}
        partner_id: {type: string}
        status: {type: string, enum: [QUEUED, RUNNING, READY, FAILED, EXPIRED]}
        provider: {type: string, enum: [local, s3]}
        object_key: {type: string}
        ciphertext_sha256: {type: string}
        ciphertext_bytes: {type: integer, format: int64}
        manifest: {type: object, additionalProperties: true}
        created_by: {type: string}
        error: {type: string}
        created_at: {type: string, format: date-time}
        completed_at: {type: [string, "null"], format: date-time}
        expires_at: {type: [string, "null"], format: date-time}
...
security:
  - sessionCookie: []